feat(nuttx): NetworkInterface PAL — getifaddrs, interface stats, resolv.conf fix
Enable GetAllNetworkInterfaces() on NuttX:
- Add pal_interfaceaddresses.c, pal_networkchange.c, pal_networkstatistics.c,
  pal_tcpstate.c to NuttX CMake build
- Bypass if_indextoname alias check on NuttX (no interface aliases)
- Undef AF_PACKET on NuttX (incomplete sockaddr_ll)
- Return zeroed stats with InterfaceUp|InterfaceHasLink from
  GetNativeIPInterfaceStatistics (NuttX has no /proc/net)
- Fix UnixIPInterfaceProperties to catch UnauthorizedAccessException
  when /etc/resolv.conf is inaccessible (NuttX has no /etc)
- Add IFF_LOOPBACK/BROADCAST/MULTICAST/ALLMULTI and _POSIX_HOST_NAME_MAX
  to nuttx-compat.h
- Enable HAVE_GETIFADDRS, HAVE_STRLCPY, HAVE_MKSTEMP, HAVE_CFMAKERAW
  in pal_config.h
- Fix ioctl casts in pal_networking.c for NuttX (unsigned long)

+++ src/native/libs/System.Native/pal_networking.c
@@ -38,6 +38,8 @@
 #include <netinet/tcp.h>
 #if HAVE_NET_IF_H
 #include <net/if.h>
+#elif defined(__NuttX__)
+#include <net/if.h>
 #endif
 #include <string.h>
 #include <sys/ioctl.h>
@@ -3016,7 +3018,7 @@ int32_t SystemNative_GetAtOutOfBandMark(intptr_t socket, int32_t* atMark)
 
     int result;
     int err;
-    while ((err = ioctl(fd, SIOCATMARK, &result)) < 0 && errno == EINTR);
+    while ((err = ioctl(fd, SIOCATMARK, (unsigned long)&result)) < 0 && errno == EINTR);
     if (err == -1)
     {
         *atMark = 0;
@@ -3041,7 +3043,7 @@ int32_t SystemNative_GetBytesAvailable(intptr_t socket, int32_t* available)
 
     int result;
     int err;
-    while ((err = ioctl(fd, FIONREAD, &result)) < 0 && errno == EINTR);
+    while ((err = ioctl(fd, FIONREAD, (unsigned long)&result)) < 0 && errno == EINTR);
     if (err == -1)
     {
         *available = 0;
